requirements.txt
click==8.1.7
colorama==0.4.6
et_xmlfile==2.0.0
Flask==2.2.3
gunicorn==20.1.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==1.23.5
openpyxl==3.1.5
pandas==1.5.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
Werkzeug==2.2.3
zipp==3.21.0
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
openpyxl==3.1.5